+++ src/LoggingCpp.cpp
@@ -1,4 +1,5 @@
 #include "../includes/LoggingCpp.h"
+#include <cmath>
 #include <cstdio>
 #include <ctime>
 #include <string>
@@ -12,4 +13,27 @@ std::string LoggingCpp::myLogger::getDate(){
 	std::string sDate = buffer;
 	return  sDate;
 } // End of myLogger::getDate
+
+std::string LoggingCpp::myLogger::getCurrTimestamp(){
+	char buffer[4096];
+	time_t uiSecond;
+	int iMs;
+	struct timespec oSpec;
+	struct tm *poNow;
+	clock_gettime(CLOCK_REALTIME, &oSpec);
+	uiSecond = oSpec.tv_sec;
+	iMs = std::round(oSpec.tv_nsec / 1.0e6);
+	if(iMs > 9999){
+		uiSecond++;
+		iMs = 0;
+	} // End of if-condition
+	
+	poNow = localtime(&uiSecond);
+	sprintf(buffer, "%04d-%02d-%02d %02d:%02d:%02d.%04d", 
+			poNow->tm_year + 1900, poNow->tm_mon, poNow->tm_mday,
+			poNow->tm_hour, poNow->tm_min, poNow->tm_sec, iMs);
+
+	std::string sTimestamp = buffer;
+	return sTimestamp;
+} // End of myLogger::getCurrTimestamp