+++ ChangeLog.md
@@ -3,7 +3,7 @@
 ## [Unrelease]
 ### [Added]
 - Add the head file `includes/LoggingCpp.h` to define the namespace and class.
-- Define the method `LoggingCpp::myLogger::getCurrentTimestamp()` to get the current timestamp.
+- Add the method `LoggingCpp::myLogger::getCurrentTimestamp()` to get the current timestamp.
 - Add the method `LoggingCpp::myLogger::getDate()` to get the current date for generating filename.
 - Define the method `LoggingCpp::myLogger::writeLog()` to write the log messages into stream.